Remove GridMap, use MeshInstances instead
This is due to it being significantly harder to set materials for each
block individually with gridmaps than with MeshInstances.
(In a gridmap, the material is set per item)
While MeshInstances are overall less performant, the hit is acceptable.
Switching back to GridMap (or similar) later always remains an option.