Add native implementation of VoxelBody
It's ugly as hell and there are some unresolved bugs with blocks not
disconnecting properly but those need a proper fix later, as the
original version also had some of these issues.

There is very heavy usage of unwrap and unsafe. The unwrap should be
replaced with some kind of custom `try_expect()?` (note ? operator) and
the unsafe code should be isolated into helper functions.