Add outline shader
It's not very efficient but I can't think of a better way to do it in
Godot.

There are 3 versions. outline_lod.shader is quite slow but it gives the
best results. outline_slow.shader is also slow and gives jagged edges.
outline_fast.shader is fast (duh) but it doesn't work nicely for small
models / big outlines.