Add Rotation/Direction struct, hide blocks surrounded by solid faces
The Rotation struct was long due. It will help with ensuring correctness
and keeping the rest of the code clean from wacky transformations based
on the rotations.

The hiding of blocks can greatly reduce the amount of vertices that
needs to be rendered. It may also speed up the mesh generation as less
things needs to be memcpyed.