Port vehicle.gd to Rust, huge overhauls
Brief summary of changed things:

  * vehicle.gd -> vehicle.rs
  * body anchors are now a tree instead of a intermingled blob.
  * Client prediction = broken & hence disabled
  * Serialization
  * Plasma cannons relative velocity

Port the GDScript vehicle implementation to Rust because GDScript is
becoming more and more of a headache. Rust is also much safer and
faster.

The body anchor system has been completely overhauled because the old
system had some quirks and complex. The current system is a tree
structure which is much easier to comprehend, use and optimize.

Client-side prediction is brokenand I'm unable to fix it. It needs to be
redesigned from scratch. To get at least some kind of multiplayer going,
the client acts will a dumb terminal for now. This introduces potentially
a lot of input latency, but what can you do...

Blocks also have been slightly redesigned to be simpler and more generic.

Serialization is now handled by serializing the full state of the
vehicle instead of original data + some state. This is a lot easier to
implement.

Plasma cannons now account for relative linear & angular velocity,
making them more realistic & preventing the projectiles from colliding
with the main body if moving too fast.